Give up on CtrlP and setup custom fuzzy finder
This function (inspired on https://github.com/garybernhardt/selecta) will use
`ag` to find all the files in the current directory, which is pretty fast, and
then perform a fuzzy find using `fyz`. The selected file will then be opened
using `:e`.

This is a lot faster than CtrlP, even without a cache (having also
the benefit that you don't need to reindex the cache every time a new file is
created). The matching algorithm that `fzy` uses is also better.

The main drawback is that it will open the file in the current buffer, so
there is not way to, for example, choose to open a file in a new tab or in a
vertical split. To do that we need different maps for each behavior:

```vimscript
" Ctrl+p opens the selected file in the current buffer
nnoremap <c-p> :call FuzzyfinderCommand(find_files_cmd, ":e")<cr>

" Ctrl+t opens the selected file in a new tab
nnoremap <c-t> :call FuzzyfinderCommand(find_files_cmd, ":tabnew")<cr>

" Ctrl+u opens the selected file in a vertical split
nnoremap <c-u> :call FuzzyfinderCommand(find_files_cmd, ":vs")<cr>
```

Also, it has a dependency on `ag` and `fzy` (`brew install ag fzy` and you're
good to go).